Use inheritance flag hints in typeinfo to optimise handler matching
One of the flags in type_info for VMI classes indicates whether there
are any bases repeated *not* by means of virtual inheritance (the
__non_diamond_repeat_mask flag).

If this is clear then we don't have to search for another matching base
after having found one (i.e. we don't have to worry about ambiguous
bases).